drivers: rename rpi_pwm_out to linux_pwm_out
