+++ cmake/configs/posix_rpi_common.cmake
@@ -78,7 +78,7 @@ set(config_module_list
 	drivers/navio_adc
 	drivers/navio_sysfs_rc_in
 	drivers/linux_gpio
-	drivers/rpi_pwm_out
+	drivers/linux_pwm_out
 	drivers/navio_rgbled
 	drivers/pwm_out_sim
 	drivers/rpi_rc_in
