+++ posix-configs/rpi/px4.config
@@ -31,6 +31,6 @@ mavlink start -d /dev/ttyUSB0
 mavlink stream -d /dev/ttyUSB0 -s HIGHRES_IMU -r 50
 mavlink stream -d /dev/ttyUSB0 -s ATTITUDE -r 50
 navio_sysfs_rc_in start
-rpi_pwm_out start
+linux_pwm_out start
 logger start -t -b 200
 mavlink boot_complete
