+++ posix-configs/rpi/px4_fw.config
@@ -29,6 +29,6 @@ mavlink stream -u 14556 -s MANUAL_CONTROL -r 10
 
 
 navio_sysfs_rc_in start
-rpi_pwm_out start -m ROMFS/px4fmu_common/mixers/AERT.main.mix
+linux_pwm_out start -m ROMFS/px4fmu_common/mixers/AERT.main.mix
 logger start -t -b 200
 mavlink boot_complete
