+++ posix-configs/rpi/px4_no_shield.config
@@ -27,6 +27,6 @@ mavlink stream -u 14556 -s ATTITUDE -r 50
 #mavlink stream -d /dev/ttyUSB0 -s HIGHRES_IMU -r 50
 #mavlink stream -d /dev/ttyUSB0 -s ATTITUDE -r 50
 rpi_rc_in start
-rpi_pwm_out start -p pca9685
+linux_pwm_out start -p pca9685
 logger start -t
 mavlink boot_complete
