+++ src/drivers/linux_pwm_out/PCA9685.h
@@ -41,7 +41,7 @@
 #include <inttypes.h>
 #include "common.h"
 
-namespace rpi_pwm_out
+namespace linux_pwm_out
 {
 
 // Register Definitions
