+++ src/drivers/linux_pwm_out/common.h
@@ -35,7 +35,7 @@
 
 #include <stdint.h>
 
-namespace rpi_pwm_out
+namespace linux_pwm_out
 {
 
 /**
