+++ src/drivers/linux_pwm_out/linux_pwm_out.cpp
@@ -55,7 +55,7 @@
 #include "navio_sysfs.h"
 #include "PCA9685.h"
 
-namespace rpi_pwm_out
+namespace linux_pwm_out
 {
 static px4_task_t _task_handle = -1;
 volatile bool _task_should_exit = false;
@@ -388,12 +388,12 @@ void usage()
 	PX4_INFO("       pwm_out status");
 }
 
-} // namespace rpi_pwm_out
+} // namespace linux_pwm_out
 
 /* driver 'main' command */
-extern "C" __EXPORT int rpi_pwm_out_main(int argc, char *argv[]);
+extern "C" __EXPORT int linux_pwm_out_main(int argc, char *argv[]);
 
-int rpi_pwm_out_main(int argc, char *argv[])
+int linux_pwm_out_main(int argc, char *argv[])
 {
 	int ch;
 	int myoptind = 1;
@@ -411,15 +411,15 @@ int rpi_pwm_out_main(int argc, char *argv[])
 	while ((ch = px4_getopt(argc, argv, "d:m:p:n:", &myoptind, &myoptarg)) != EOF) {
 		switch (ch) {
 		case 'd':
-			strncpy(rpi_pwm_out::_device, myoptarg, sizeof(rpi_pwm_out::_device));
+			strncpy(linux_pwm_out::_device, myoptarg, sizeof(linux_pwm_out::_device));
 			break;
 
 		case 'm':
-			strncpy(rpi_pwm_out::_mixer_filename, myoptarg, sizeof(rpi_pwm_out::_mixer_filename));
+			strncpy(linux_pwm_out::_mixer_filename, myoptarg, sizeof(linux_pwm_out::_mixer_filename));
 			break;
 
 		case 'p':
-			strncpy(rpi_pwm_out::_protocol, myoptarg, sizeof(rpi_pwm_out::_protocol));
+			strncpy(linux_pwm_out::_protocol, myoptarg, sizeof(linux_pwm_out::_protocol));
 			break;
 
 		case 'n': {
@@ -433,44 +433,44 @@ int rpi_pwm_out_main(int argc, char *argv[])
 					max_num = actuator_outputs_s::NUM_ACTUATOR_OUTPUTS;
 				}
 
-				rpi_pwm_out::_max_mum_outputs = max_num;
+				linux_pwm_out::_max_mum_outputs = max_num;
 			}
 			break;
 		}
 	}
 
 	// gets the parameters for the esc's pwm
-	param_get(param_find("PWM_DISARMED"), &rpi_pwm_out::_pwm_disarmed);
-	param_get(param_find("PWM_MIN"), &rpi_pwm_out::_pwm_min);
-	param_get(param_find("PWM_MAX"), &rpi_pwm_out::_pwm_max);
+	param_get(param_find("PWM_DISARMED"), &linux_pwm_out::_pwm_disarmed);
+	param_get(param_find("PWM_MIN"), &linux_pwm_out::_pwm_min);
+	param_get(param_find("PWM_MAX"), &linux_pwm_out::_pwm_max);
 
 	/*
 	 * Start/load the driver.
 	 */
 	if (!strcmp(verb, "start")) {
-		if (rpi_pwm_out::_is_running) {
+		if (linux_pwm_out::_is_running) {
 			PX4_WARN("pwm_out already running");
 			return 1;
 		}
 
-		rpi_pwm_out::start();
+		linux_pwm_out::start();
 	}
 
 	else if (!strcmp(verb, "stop")) {
-		if (!rpi_pwm_out::_is_running) {
+		if (!linux_pwm_out::_is_running) {
 			PX4_WARN("pwm_out is not running");
 			return 1;
 		}
 
-		rpi_pwm_out::stop();
+		linux_pwm_out::stop();
 	}
 
 	else if (!strcmp(verb, "status")) {
-		PX4_WARN("pwm_out is %s", rpi_pwm_out::_is_running ? "running" : "not running");
+		PX4_WARN("pwm_out is %s", linux_pwm_out::_is_running ? "running" : "not running");
 		return 0;
 
 	} else {
-		rpi_pwm_out::usage();
+		linux_pwm_out::usage();
 		return 1;
 	}
